Move websocket dependencies to module
